requirements.txt
django~=5.2
channels~=4.3.2
django-cors-headers
uvicorn
drf-yasg~=1.21.11
pyyaml~=6.0.3
asgiref~=3.11.0
djangorestframework~=3.16.1